requirements.txt
discord.py
aiohttp
asyncpg
python-dotenv
flask
SQLAlchemy
